popping star display for ratings into its own component
