+++ src/components/RatingStars.js
@@ -1,0 +1,36 @@
+import React, {useState, useEffect} from 'react';
+import {Row, Alert} from 'react-bootstrap';
+import { AiOutlineStar, AiFillStar } from "react-icons/ai";
+
+
+const RatingStars = ({currentRating, maxRating}) => {
+
+    const displayRating = () => {
+        const ratingIcons = []
+        console.log('entered displayRating')
+
+        for(let i = 0; i < currentRating; i++){
+            console.log('adding a star')
+            ratingIcons.push(
+                <AiFillStar/> 
+            )
+        }
+        if (currentRating < maxRating) {
+            for(let i = currentRating; i < maxRating; i++){
+                console.log('adding a star')
+                ratingIcons.push(
+                    <AiOutlineStar/> 
+                )
+            }
+        }
+        return (ratingIcons.map((icon) => icon))
+    }
+
+    return (
+        <Row className='justify-content-center'>
+            {displayRating()}
+        </Row>
+    )
+}
+
+export default RatingStars;
